with logo

+++ resources/views/layouts/app.blade.php
@@ -70,15 +70,15 @@
 @yield('js')
 <script  type="text/javascript">
     console.log('dfdddd');
-    function clearLocalStorage() {
-        console.log('dfd');
+    if(document.getElementById('guest')) {
+        console.log('dfdf');
         localStorage.clear();
     }
 </script>
 <body>  
     <div id="app" class="container-fluid">
         @guest
-        <div class=" row d-md-flex h-md-100 align-items-center" onload="clearLocalStorage()">
+        <div class=" row d-md-flex h-md-100 align-items-center" id="guest">
 
             <!-- First Half -->
